requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
mysqlclient==2.2.0
requests==2.31.0
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.0
